pyproject.toml
[tool.poetry]
name =  "databricks_genai_hackt"
version = "0.1.0"
description = "This Git repository contains the source code and the Python package for the project heb_databricks_genai."
authors = ["Rajesh Chodavarapu <chodavarapu.rajesh@heb.com>" ]
readme = "README.md"
packages = [
    { include = "databricks_genai_hackt" },
]
homepage = "https://github.com/chorajesh/databricks_genai_hackt"
repository = "https://github.com/chorajesh/databricks_genai_hackt"

[tool.poetry.dependencies]
python = "3.11.4"
python-dotenv = "^1.0.0"
openai = "0.28"
streamlit = "1.30.0"
streamlit-chat = "0.1.1"
transformers  = "4.37.1"
databricks-genai-inference = "0.1.3"
databricks-vectorsearch = "0.22"
torch = "2.1.2"
sentence-transformers = "2.3.1"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
autoflake8 = "^0.4.1"
flake8 = "^7.0.0"
coverage = "^7.4.0"
mypy = "^1.8.0"
isort = "^5.13.2"
autoflake = "^2.2.1"
pytest = "^7.4.4"
pytest-cov = "^4.1.0"
ipykernel = "^6.29.0"

[tool.mypy]
python_version = ">=3.11"
disallow_any_generics = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_untyped_decorators = true
ignore_errors = false
ignore_missing_imports = true
implicit_reexport = false
strict_optional = true
strict_equality = true
no_implicit_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unreachable = true
warn_no_return = true
warn_return_any = true

[tool.flake8]
max-line-length = 120
exclude = ".git __pycache__ .mypy_cache .pytest_cache .reports notebooks _resource"
extend-ignore = ["E203"]

[tool.pytest.ini_options]
addopts = "--pdbcls=IPython.terminal.debugger:TerminalPdb -p no:warnings"
junit_family = "xunit2"
norecursedirs = ".git __pycache__ .mypy_cache .pytest_cache .reports notebooks _resource"

[tool.coverage.report]
omit = ["*/__init__.py", ".mypy_cache", ".reports", "notebooks"]
show_missing = true

[tool.black]
line-length = 120
target-version = ["py311"]
exclude = '(\.git|\.mypy_cache|\.pytest_cache|\.reports|build|dist|notebooks|_resource)'

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
main = "heb_databricks_genai.__main__:main"
streamlit_ui = "recipe_chatbot.__main__:run_chatbot_streamlit_ui"

